requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.2
black==23.3.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
coverage==7.2.3
cssselect==1.2.0
cycler==0.11.0
distlib==0.3.6
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
dnspython==2.3.0
docker==6.0.1
duckduckgo-search==2.8.5
exceptiongroup==1.1.1
filelock==3.11.0
flake8==6.0.0
fonttools==4.39.4
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
gTTS==2.3.1
h11==0.14.0
html5lib==1.1
httplib2==0.22.0
identify==2.5.22
idna==3.4
iniconfig==2.0.0
isort==5.12.0
kiwisolver==1.4.4
loguru==0.7.0
lxml==4.9.2
matplotlib==3.7.1
mccabe==0.7.0
mpl-finance==0.10.1
mplfinance==0.12.9b7
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nodeenv==1.7.0
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.2
orjson==3.8.10
outcome==1.2.0
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==3.2.0
playsound==1.2.2
pluggy==1.0.0
pre-commit==3.2.2
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.3.1
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
PyYAML==6.0
readability-lxml==0.8.1
redis==4.5.4
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-fuzzy==0.4.2
scipy==1.10.1
selenium==4.8.3
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sourcery==1.2.0
sqlparse==0.4.4
tiktoken==0.3.3
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.13.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.21.0
webdriver-manager==3.8.6
webencodings==0.5.1
websocket-client==1.5.1
win32-setctime==1.1.0
wsproto==1.2.0
yarl==1.8.2
